wsd: Announce structure made exported
Announce is the common payload for Hello, ProbeMatches and ResolveMatches
messages.

As these messages have very similar semantics, it is quite convenient to
have a common type that can represent their content. It makes possible to
implement a common handling code.

Also added the methods to export payload of these messages as slice
of Announce structures:

    func (Hello) Announces() []Announce
    func (ProbeMatches) Announces() []Announce
    func (ResolveMatches) Announces() []Announce