cpython: Python->Go value conversions revisited
The single universal Object.Unbox doesn't play well with mapping
Python type system to Go type system.

There is just too many questions hard to answer. For example,
should we automatically convert Python dictionaries to the Go maps?
If Object represents a container type, and value is updated
at the Go side, should converted value be updates as well?
And so on.

So now we have an assortment of the Object.Bool, Object.Int, Object.Float
and so on methods that performs conversion on demand and into the
explicitly requested type.