scriptable: added go->protocol mapping of keywords
At the Go side, keyword names (say, field names of the
escl.ScannerCapabilities structure) are automatically obtained
from the Go symbol names.

These names are aimed to 1:1 match the corresponding protocol names, but
unfortunately, sometimes golint dictates us a different spelling (for example,
it recognizes the acronym Uuid and dictates it to be spelled as UUID).

So when exposing these names to the model/protocol side, some mapping
is required...