argv: if there are long options, "--" at the end offers them for completion
The "--" at the end is ambiguous in the context of completion: it can
be interpreted either ass end of the options list of as a beginning of
the long option.

With this change, if command has long options, "--" at the end of the
command line interpreted as the beginning of the option name. Otherwise,
it is interpreted as the end of the option list.