transport: added ServerQuery type
ServerQuery maintains the HTTP Server query processing,
allowing per-request centralized logging, hooking and so on.

It binds together http.Request, the corresponding http.ResponseWrapper
and provides more clean and convenient HTTP request processing context,
that the stdlib http package does.

It was creates as the internal part of the proto/escl package, and
now slightly refactored and exported.