usbip: the big refactoring
  - support for multiple virtual USB devices
  - asynchronous (non-blocking) execution of the USBIP_CMD_SUBMIT requests
  - support for USBIP_CMD_UNLINK requests
  - full support of strings in the USB descriptors
  - virtual IPP over USB printer integrated with the http.Server
  - symbols that are not part of the public API made unexported

Handling of the configuration file removed for now and will be
replaced by the integration with the common GO-MFP infrastructure.

Support of the outgoing IPP connections temporary removed and
will be replaced with the mfp-proxy/mfp-virtual integration, which
gives us a first class support of the IPP over USB simulation
and proxying.