log: added new backend and loggers
The newly added Stderr backend writes logs to the stderr. Unlike the
Console backend, it always use stderr and don't use terminal colors.

The newly added FatalLogger logger writes LevelFatal logs to the stderr
and discards less important messages.

The newly added StderrLogger writes logs to the stderr and ConsoleLogger
logger writes logs to the console.

The DefaultLogger redefined as the alias to the ConsoleLogger.

Logging via context.Context uses FatalLogger instead of the DiscardLogger
if the context has no associated logger.