usbip: EndpointConn and EndpointListener made exported types.
EndpointConn implements the net.Conn interface on a top of the Endpoint.

EndpointListener implements the net.Listener interface on a top of
the group of Endpoints.

These types a suitable for usage together with the http.Server in
order to implement the HTTP part of the IPP over USB protocol or
for similar tasks.