cpython: Object.Int, Uint, Float, and Complex support automatic conversion
The following conversions are now possible:

    Object.Int and Object.Uint  will accept integer and floating point values
    Object.Float will accept integer and floating point values
    Object.Complex will accept integer, floating point and complex values

Conversion from the floating point to integer performed by discarding
the fractional part regardless of the floating point value sing, which
is consistent with the Go specification.

Range errors are checked and reported (so huge float will not silently
convert into int that can't handle its value).