usbip, usbhost: fixed setup packet handling for printer class-specific requests
wValue and wIndex fields of the USB setup packets for the printer
class-specific control requests contain zero-based configuration,
interface and alt setting indices, not bConfigurationValue,
bInterfaceNumber and bAlternateSetting as it was previously implemented
due to the specification misinterpretation.

Now we have a consensus with the CUPS usb backend regarding IEEE-1284
Device ID access! :)