ipp: PrinterAttributes rearranged a bit
MediaCol and friends moved from job-attrs.go into media-attrs.go.

Added MediaColDatabase struct and embedded into the PrinterAttributes
instead of the direct inclusion of the MediaColDatabase field.

The final goal is to make XXX.KnownAttrs() the authoritative source
of list of attributes included into the particular group, when executing
Get-Printer-Attributes and similar operations.