ipp-registrations-generate: semantic of <link> in errata.xml changed
Previously, <link> was used to resolve ambiguous or broken links
by providing the absolute path to the target. Now we have the much
more convenient <subst> element for this purpose.

So now <link> injects the missed link:

  <link>
     <from>Document Status/document-format-details(deprecated)</from>
     <from>Document Status/document-format-details-detected(deprecated)</from>
     <from>Job Status/document-format-details-detected(deprecated)</from>
     <from>Job Status/document-format-details-supplied</from>
     <to>Operation/document-format-details</to>
   </link>

Note, multiple <from> elements are allowed. The <to> element may be either
absolute or relative. Relative <to> elements are resolved the usual way.