ipp: refactoring
New parser for the "ipp:" struct fields was written. This parser
understands IANA notation for attribute syntax and returns parsed
syntax as *iana.DefAttr structure.

This work is one of the step required to integrate IPP encoder/decoder
with the database IANA registrations for IPP.

Note, the new parser was test-covered but not yet integrated.